lib: Make build backend hooks reusable
It will be used by other parts of this project.

Fixes: https://github.com/stanislavlevin/pyproject_installer/issues/40
